+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## [0.2.0] - 2020-05-07
+* Make adaptive_scaffold completely compatible with 1.7.0
+* Add a bottom_navbar_theme until flutter 1.18.0-pre 6 (which contains the Material implementation for BottomNavigationBar) reaches stable
+
 ## [0.1.0] - 2020-05-07
 
 * Initial Release.